[gpx] Move MapPredefinedColor to color parser Signed-off-by: cyber-toad <the.cyber.toad@proton.me>
