+++ libs/kml/color_parser.hpp
@@ -4,6 +4,8 @@
 #include <optional>
 #include <string_view>
 
+#include "types.hpp"
+
 namespace kml
 {
 
@@ -17,5 +19,8 @@ constexpr uint32_t ToRGBA(Channel red, Channel green, Channel blue, Channel alph
 std::optional<uint32_t> ParseHexColor(std::string_view c);
 std::optional<uint32_t> ParseGarminColor(std::string_view c);
 std::optional<uint32_t> ParseOSMColor(std::string_view c);
+
+PredefinedColor MapPredefinedColor(uint32_t rgba);
+std::string_view MapGarminColor(uint32_t rgba);
 
 }  // namespace kml
